fix: use async generateKeyPair
The previous implementation wrapped a synchronous function in a promise.

This commit uses the async version 'generateKeyPair' to make it a proper
async function. This means if in future we use an encryption algorithm
that takes longer e.g. 'rsa', the thread will not be blocked as the
encryption process runs.