feat: add user journey choice
This commit implements a simple TUI to choose the user journey; 'read' |
'write'. It also adds the getTimeStamp function to create more readable
timestamps when choosing an entry on the 'read' journey.

The TUI uses Node's readline.emitKeypressEvents method with stdin/out
which allows us to react to individual key presses. This provides
immediate feedback and re-rendering to be possible when using 'j' & 'k'
for journal entry list navigation.